refresh bar new working.
