+++ index.html
@@ -111,8 +111,8 @@
         <div class="onoffswitch">
           <input type="checkbox" name="onoffswitch" class="onoffswitch-checkbox" id="myonoffswitch">
           <label class="onoffswitch-label" for="myonoffswitch">
-              <span class="onoffswitch-inner"></span>
-              <span class="onoffswitch-switch"></span>
+            <span class="onoffswitch-inner"></span>
+            <span class="onoffswitch-switch"></span>
           </label>
         </div>
 
@@ -127,6 +127,11 @@
           </div>
         </div>
         <!-- DROPDOWN ENDS -->
+        
+        <!-- REFRESH BUTTON -->
+        <p id="refresh-button">
+          <img src="images/refreshbutton.png" alt="Refresh Button" title="Refresh DB" width="30px" height="30px">
+        </p>
 
       </div>
 
@@ -344,17 +349,14 @@
 
 
 	  <div class="row" id="connectwithus">
-		<h2 class="text-center d-block">Connect with us</h2>
-		<!-- <p>You can connect with us on any of these platforms:</p> -->
-	  <div id="connectwithusbody">
-		  <a class="col-md-2 links " href="https://www.facebook.com/PublicLab/" title="Facebook" ><i class="fa fa-facebook fa-3x"></i> </a>
-		  <a class="col-md-2 links " href="https://twitter.com/PublicLab?ref_src=twsrc%5Egoogle%7Ctwcamp%5Eserp%7Ctwgr%5Eauthor" title="Twitter" ><i class="fa fa-twitter fa-3x"></i> </a>
-		  <a class="col-md-2 links " href="https://github.com/publiclab" title="Organization" ><i class="fa fa-github fa-3x"></i> </a>
-      		  <a class="col-md-2 links " href="https://publiclab.org" title="Publiclab Website"><i class="fa fa-globe fa-3x"></i></a>
-                  <a class="col-md-2 links " href="https://gitter.im/publiclab/publiclab" title="Gitter"><i class="fa fa-comment fa-3x"></i></a>
-	  </div>
-	  </div>
-    </!-->
+      <h2 class="text-center d-block">Connect with us</h2>
+      <div id="connectwithusbody">
+        <a class="col-md-2 links " href="https://www.facebook.com/PublicLab/" title="Facebook" ><i class="fa fa-facebook fa-3x"></i> </a>
+        <a class="col-md-2 links " href="https://twitter.com/PublicLab?ref_src=twsrc%5Egoogle%7Ctwcamp%5Eserp%7Ctwgr%5Eauthor" title="Twitter" ><i class="fa fa-twitter fa-3x"></i> </a>
+        <a class="col-md-2 links " href="https://github.com/publiclab" title="Organization" ><i class="fa fa-github fa-3x"></i> </a>
+        <a class="col-md-2 links " href="https://publiclab.org" title="Publiclab Website"><i class="fa fa-globe fa-3x"></i></a>
+      </div>
+    </div>
 
     <footer>
       <div class="row">
